backend/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
djangorestframework==3.15.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-env==1.0.0
sqlparse==0.5.0
